Relocate Exit_Interview classes
