+++ src/Telemetry/Exit_Interview/Exit_Interview_Subscriber.php
@@ -6,7 +6,7 @@
  *
  * @package StellarWP\Telemetry
  */
-namespace StellarWP\Telemetry;
+namespace StellarWP\Telemetry\Exit_Interview;
 
 use StellarWP\Telemetry\Contracts\Abstract_Subscriber;
